+++ packages/api-jwt/src/lib/routes/authorize/refresh.ts
@@ -25,16 +25,14 @@ export class PluginRoute extends Route {
 		const tokenData = await this.container.jwt.decrypt<RESTPostOAuth2AccessTokenResult>(refresh_token, 'refresh_token');
 		if (isNullish(tokenData)) return response.error('Could not get the user information in the token.');
 		if (isNullish(tokenData.refresh_token)) {
-			this.container.logger.warn(
-				'The access token cannot be refreshed because the strategy for obtaining persistent sessions was not assigned.'
-			);
+			this.container.logger.warn('The access token cannot be refreshed because the hook for get persistent sessions was not assigned.');
 
 			return response.error('Persistent sessions not active.');
 		}
 
-		const refreshData = await this.container.jwt.auth(tokenData.refresh_token, 'refresh');
+		const refreshData = await this.container.jwt.auth(tokenData.data.refresh_token, 'refresh');
 		if (isNullish(refreshData)) return response.error('There was a problem refreshing data in Discord.');
 
-		return response.status(200).json(this.container.jwt.encrypt(refreshData));
+		return response.status(200).json(await this.container.jwt.encrypt(refreshData));
 	}
 }