+++ packages/env/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.4...@kaname-png/plugin-env@1.1.5) (2022-12-29)
+
+**Note:** Version bump only for package @kaname-png/plugin-env
+
+## [1.1.4](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.3...@kaname-png/plugin-env@1.1.4) (2022-12-29)
+
+### Bug Fixes
+
+-   import typing resolution ([41f47ff](https://github.com/kaname-png/neko-plugins/commit/41f47ffc58d8b8ebe4a06804ed736eda7f19f12a))
+
 ## [1.1.3](https://github.com/kaname-png/neko-plugins/compare/@kaname-png/plugin-env@1.1.2...@kaname-png/plugin-env@1.1.3) (2022-12-04)
 
 **Note:** Version bump only for package @kaname-png/plugin-env